OpenVX HAL updated to use generic OpenVX wrappers
